Remove a citation placeholder that's nigh-impossible to fill.
To the best of our knowledge, all studies that group species have grouped
them the same in their competitive effects and responses. This is
nigh-impossible to provide a citation for, as proving it by citation
is an attempt to prove a negative. One cannot simply cite a bunch of studies
that have made this assumption, but I cannot find a paper that even
identifies that the assumptions is typically made.