Adapt depend.sh to use $(from_root) to be agnostic of where it is run.
Previously, running make in this repo or in a parent repo would generate
different .d files, where the value of $(from_root) was hard-coded.
Hence, while make could be run either from this repo or a parent repo,
it always had to be run from the *same* repo.
Running make in one repo, then in another, would break on incompatible .d files.
This fixes that.