Change the post-processing to not store parameter objects in a data frame.
Storing R6 objects in one column of a data frame inevitably played havoc
with the cache when averaging them, not to mention the cost to the cache
of converting each row to an object in the first place. This change
ensures they remain stored in a flat data frame, and are only converted
to an R6 object as necessary.