Swap the order of filenames in the arguments to the C++ program.
Now, the output file followed by the three input files, rather than vice versa.
As the set of input files begin to vary (with invidual response vs time series),
and with the prior files, this will put the constant arguments at the beginning.
This now also more closely matches the arguments to the guess-priors script.