Make Solver::getGroupings() an abstract function.
For historical reason, Solver::getGrouping() was an abstract function
specialised by MaximumLikelihoodSolver and ReversibleJumpSolver,
but getGroupings() was implemented separately by the two.
That's clearly not appropriate, so now getGroupings() is an abstract
function overridden by each of the two, while getGrouping() has
a unified, non-abstract implementation in Solver, using getGroupings().