Remove the virtual Solver::getGrouping(), defining it separately in each subclass.
It made sense in a MaximumLikelihoodSolver, but not in a
ReversibleJumpSolver, where there is both currentGroupings
and proposedGroupings, and it must return one or other as
appropriate. This leads to a lot of funky logic, where certain
things have to be calculated before or after proposing a jump,
and is just a pain all around. Defining them separately should allow
the logic throughout ReversibleJumpSolver to be cleaned up a lot.