Change dialing in to look at all parameters, not just the first.
It used just the first growth rate and the first competition coefficient,
for simplicity. This could lead to bias if the first species had
a drastically different growth rate or competitive effect/response
than other species. So use them all instead. This requires ungrouping them,
finding the variance of each of them, and then averaging that variance.