Have time series data ignore species not present at the start of the time span.
Should've done this in the first place, but it didn't occur to me.
It will have been messing with the error variances, as there
will have been a *lot* of zero errors, due to correctly predicting
zero abundance for absent species. This won't have affected
likelihood ratios, but may have had other effects. Hopefully,
the code will also be faster with not having to solve for a lot
of zeroes with every solveIVP() call.