Don't change the trans-model jump variance multiplier.
Changing it was just messing with things. Having it equal to the
within-model jump variance multipler dragged it too low. Trying to
force a trans-model jump acceptance rate of 0.23 just didn't work;
it would never go go that high. Trying to balance merge and split
acceptance rates did very different and weird things to the
coclassification matrices for each dataset. And throughout all the
changes, it never really made all that much difference to the
trans-model jump acceptance rates. It seems to go best if I just
leave it fixed at 1.