Explicitly extract options in the article analysis scripts, without attach().
The attach() function doesn't create any objects in the parent
environment, it simply adds teh attached object to the search path.
As such, when we save the data with save(list = ls()), the
"attachment" is lost. As such, while "max_rhat", read as an option,
works within the article-analysis-2 script, as the options have been
attached, it doesn't work within an article file that loads the
saved .rda file. As such, we assign the options into objects in the
global environment properly, so that they are saved and loaded.