Move the clean rule from Makefile to Rules.mk.
It defined simply as having the prerequisite of the
$(from_root)-clean rule, and so remains compatible with any clean
rules defined at higher levels by makefiles that include this one.
It is more elegant this way, and requires less boilerplate in
this makefile, and in higher makefiles that include this one.
It does have the drawback that higher makefiles can't define a
clean rule that *doesn't* clean this project, but that's acceptable.