Replace ReversibleJumpSolverBase with ReversibleJumpSolverInterface.
Previously, ReversibleJumpSolver inherited from ReversibleJumpSolverBase,
which in turn inherited from Solver. Now, ReversibleJumpSolver inherits
from ReversibleJumpSolverInterface and from Solver. This will allow Solver
itself (or something like it) to be templated on the error distribution,
as ReversibleJumpSolver is. This will be necessary when a
MaximumLikelihoodSolver will also be able to vary in error distribution.