Use NLopt's ftol instead of xtol for termination.
That is, base termination on changes in log likelihood, rather
than parameter values. This still seems to optimise very effectively,
yet runs much faster with the same relative tolerance (because,
as explained in the NLopt documentation, a given xtol is
approximately equal to an ftol about the square of that).