Return once more to using ftol for NLopt termination.
In the group resolution test, we're getting parameter identifiability
issues, where the parameters for a given row of the alpha matrix
(in proportion to the corresponding growth rate) is exploding,
taking rapidly growing values while very marginally improving the
log likelihood. Setting the tolerance based on the log likelihood,
rather than based on the parameter values themselves, should allow
these cases to terminate in a more reasonable time.