Set tolerances of the NLopt solver from the command line.
There does not to be a single xtol/ftol that works for both the
true TCL data and the group resolution test. Using xtol = 1e-3 works
fine for the true data. Using this for the group resolution data causes
non-termination due to parameter identifiability issues. Meanwhile,
using ftol = 5e-5 seems to work for the group resolution tests,
but does not sufficiently good fits for the true data. As such,
by setting them from the command line, we can use appropriate values
for each.